+++ docs/changelog.md
@@ -73,6 +73,8 @@
     * Pyroberry Bush
         * Spawns in desert
         * Drops Pyroberries when sheared at max growth stage
+    * Pyroberry Wine
+    * Pyroberry Cocktail
     * Ore Compat:
         * Bloom yield and slag bloom yield can now be adjusted in the ore compat json files [(#310)](https://github.com/codetaylor/pyrotech-1.12/issues/310)
     * Config option to:
@@ -295,6 +297,7 @@
         * `fluid.pyroberry_wine`
         * `tile.fluid.pyroberry_wine.name`
         * `item.pyrotech.pyroberry.wine.name`
+        * `item.pyrotech.pyroberry.cocktail.name`
 
     * **Changed Lang Keys:**
         * `tile.pyrotech.stone_bricks.name` -> `tile.pyrotech.masonry_brick_block.name`